requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.6.1
django-cors-headers==4.7.0
django-js-asset==3.1.0
django-simpleui==2025.1.13
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
